+++ sdk/kubernetesconfiguration/arm-kubernetesconfiguration/src/sourceControlConfigurationClient.ts
@@ -8,11 +8,6 @@
 
 import * as coreClient from "@azure/core-client";
 import * as coreRestPipeline from "@azure/core-rest-pipeline";
-import {
-  PipelineRequest,
-  PipelineResponse,
-  SendRequest
-} from "@azure/core-rest-pipeline";
 import * as coreAuth from "@azure/core-auth";
 import {
   ExtensionsImpl,
@@ -20,7 +15,7 @@ import {
   FluxConfigurationsImpl,
   FluxConfigOperationStatusImpl,
   SourceControlConfigurationsImpl,
-  OperationsImpl
+  OperationsImpl,
 } from "./operations";
 import {
   Extensions,
@@ -28,14 +23,13 @@ import {
   FluxConfigurations,
   FluxConfigOperationStatus,
   SourceControlConfigurations,
-  Operations
+  Operations,
 } from "./operationsInterfaces";
 import { SourceControlConfigurationClientOptionalParams } from "./models";
 
 export class SourceControlConfigurationClient extends coreClient.ServiceClient {
   $host: string;
   subscriptionId: string;
-  apiVersion: string;
 
   /**
    * Initializes a new instance of the SourceControlConfigurationClient class.
@@ -46,7 +40,7 @@ export class SourceControlConfigurationClient extends coreClient.ServiceClient {
   constructor(
     credentials: coreAuth.TokenCredential,
     subscriptionId: string,
-    options?: SourceControlConfigurationClientOptionalParams
+    options?: SourceControlConfigurationClientOptionalParams,
   ) {
     if (credentials === undefined) {
       throw new Error("'credentials' cannot be null");
@@ -61,10 +55,10 @@ export class SourceControlConfigurationClient extends coreClient.ServiceClient {
     }
     const defaults: SourceControlConfigurationClientOptionalParams = {
       requestContentType: "application/json; charset=utf-8",
-      credential: credentials
+      credential: credentials,
     };
 
-    const packageDetails = `azsdk-js-arm-kubernetesconfiguration/6.1.1`;
+    const packageDetails = `azsdk-js-arm-kubernetesconfiguration/7.0.0-beta.1`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
@@ -74,20 +68,21 @@ export class SourceControlConfigurationClient extends coreClient.ServiceClient {
       ...defaults,
       ...options,
       userAgentOptions: {
-        userAgentPrefix
+        userAgentPrefix,
       },
       endpoint:
-        options.endpoint ?? options.baseUri ?? "https://management.azure.com"
+        options.endpoint ?? options.baseUri ?? "https://management.azure.com",
     };
     super(optionsWithDefaults);
 
     let bearerTokenAuthenticationPolicyFound: boolean = false;
     if (options?.pipeline && options.pipeline.getOrderedPolicies().length > 0) {
-      const pipelinePolicies: coreRestPipeline.PipelinePolicy[] = options.pipeline.getOrderedPolicies();
+      const pipelinePolicies: coreRestPipeline.PipelinePolicy[] =
+        options.pipeline.getOrderedPolicies();
       bearerTokenAuthenticationPolicyFound = pipelinePolicies.some(
         (pipelinePolicy) =>
           pipelinePolicy.name ===
-          coreRestPipeline.bearerTokenAuthenticationPolicyName
+          coreRestPipeline.bearerTokenAuthenticationPolicyName,
       );
     }
     if (
@@ -97,7 +92,7 @@ export class SourceControlConfigurationClient extends coreClient.ServiceClient {
       !bearerTokenAuthenticationPolicyFound
     ) {
       this.pipeline.removePolicy({
-        name: coreRestPipeline.bearerTokenAuthenticationPolicyName
+        name: coreRestPipeline.bearerTokenAuthenticationPolicyName,
       });
       this.pipeline.addPolicy(
         coreRestPipeline.bearerTokenAuthenticationPolicy({
@@ -107,9 +102,9 @@ export class SourceControlConfigurationClient extends coreClient.ServiceClient {
             `${optionsWithDefaults.endpoint}/.default`,
           challengeCallbacks: {
             authorizeRequestOnChallenge:
-              coreClient.authorizeRequestOnClaimChallenge
-          }
-        })
+              coreClient.authorizeRequestOnClaimChallenge,
+          },
+        }),
       );
     }
     // Parameter assignments
@@ -117,44 +112,14 @@ export class SourceControlConfigurationClient extends coreClient.ServiceClient {
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.apiVersion = options.apiVersion || "2023-05-01";
     this.extensions = new ExtensionsImpl(this);
     this.operationStatus = new OperationStatusImpl(this);
     this.fluxConfigurations = new FluxConfigurationsImpl(this);
     this.fluxConfigOperationStatus = new FluxConfigOperationStatusImpl(this);
     this.sourceControlConfigurations = new SourceControlConfigurationsImpl(
-      this
+      this,
     );
     this.operations = new OperationsImpl(this);
-    this.addCustomApiVersionPolicy(options.apiVersion);
-  }
-
-  /** A function that adds a policy that sets the api-version (or equivalent) to reflect the library version. */
-  private addCustomApiVersionPolicy(apiVersion?: string) {
-    if (!apiVersion) {
-      return;
-    }
-    const apiVersionPolicy = {
-      name: "CustomApiVersionPolicy",
-      async sendRequest(
-        request: PipelineRequest,
-        next: SendRequest
-      ): Promise<PipelineResponse> {
-        const param = request.url.split("?");
-        if (param.length > 1) {
-          const newParams = param[1].split("&").map((item) => {
-            if (item.indexOf("api-version") > -1) {
-              return "api-version=" + apiVersion;
-            } else {
-              return item;
-            }
-          });
-          request.url = param[0] + "?" + newParams.join("&");
-        }
-        return next(request);
-      }
-    };
-    this.pipeline.addPolicy(apiVersionPolicy);
   }
 
   extensions: Extensions;