+++ sdk/kubernetesconfiguration/arm-kubernetesconfiguration/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Release History
+    
+## 7.0.0-beta.1 (2024-04-08)
+    
+**Features**
 
+  - Added Interface LayerSelectorDefinition
+  - Added Interface LayerSelectorPatchDefinition
+  - Added Interface MatchOidcIdentityDefinition
+  - Added Interface MatchOidcIdentityPatchDefinition
+  - Added Interface OCIRepositoryDefinition
+  - Added Interface OCIRepositoryPatchDefinition
+  - Added Interface OCIRepositoryRefDefinition
+  - Added Interface OCIRepositoryRefPatchDefinition
+  - Added Interface PostBuildPatchDefinition
+  - Added Interface SubstituteFromPatchDefinition
+  - Added Interface TlsConfigDefinition
+  - Added Interface TlsConfigPatchDefinition
+  - Added Interface VerifyDefinition
+  - Added Interface VerifyPatchDefinition
+  - Added Type Alias OperationType
+  - Interface FluxConfiguration has a new optional parameter ociRepository
+  - Interface FluxConfigurationPatch has a new optional parameter ociRepository
+  - Added Enum KnownOperationType
+  - Enum KnownSourceKindType has a new value OCIRepository
+
+**Breaking Changes**
+
+  - Interface SourceControlConfigurationClientOptionalParams no longer has parameter apiVersion
+  - Type of parameter postBuild of interface KustomizationPatchDefinition is changed from PostBuildDefinition to PostBuildPatchDefinition
+  - Class SourceControlConfigurationClient no longer has parameter apiVersion
+    
 ## 6.1.1 (Unreleased)
 
 ### Features Added